initial commit: This is a file I made to test out analyses and create some functions. So far functions are: + getURL: Used to safely query online sources + imdb_downloader: lets the user easily download data from imbd without permanetly downloading the file.
Other thoughts:
I'm attemptong to get the data ready for the simple James Bond Analysis but I am discovering some James Bond movies are missing? Also the way that the current data structure is set up, it would be possible to make this a shiny app, but we probably would not be able to upload it, due to the size of these imdb dataframes.